feat: refactor Home page structure and add new sections for improved layout and functionality
